+++ R/RcppExports.R
@@ -13,8 +13,8 @@ indep_test_cpp <- function(x, y, method, weights, remove_missing, alternative) {
     .Call(`_wdm_indep_test_cpp`, x, y, method, weights, remove_missing, alternative)
 }
 
-rank_wtd <- function(x, weights, ties_method = "min") {
-    .Call(`_wdm_rank_wtd`, x, weights, ties_method)
+rank_wtd_cpp <- function(x, weights, ties_method = "min") {
+    .Call(`_wdm_rank_wtd_cpp`, x, weights, ties_method)
 }
 
 perm_sum_cpp <- function(x, k) {